MissionControl (MissionControl v1.0.0)

Documentation for MissionControl. MissionControl calculate fuel required for the flight: launch from one planet of the Solar system, and to land on another planet of the Solar system, depending on the flight route.

Functions

calculate(mass, stages)

@spec calculate(non_neg_integer(), [tuple()]) :: Integer.t() | String.t()

MissionControl.calculate/2

Calculate fuel required for whole Mission: Launch and Land on each Planet in list

Examples

iex> MissionControl.calculate(28801, [{:launch, 9.807}, {:land, 1.62}, {:launch, 1.62}, {:land, 9.807}])
51898

calculate(mass, stage, planet)

@spec calculate(non_neg_integer(), atom(), float()) :: Integer.t() | String.t()

MissionControl.calculate/3

Calculate fuel required to perform single Stage of Mission: Launch or Land

Examples

iex>  MissionControl.calculate(28801, :land, 9.807)
13447

iex>  MissionControl.calculate(28801, :land, :earth)
13447

help()

@spec help() :: %{
  examples: [{:mission, <<_::808>>} | {:mission_stage, <<_::760>>}, ...],
  methods: [{:mission, <<_::368>>} | {:mission_stage, <<_::416>>}, ...],
  requirenments: [
    {:mission_stages, [...]} | {:planets_available, [...]} | {:weight, [...]},
    ...
  ]
}

MissionControl.help()

Returns help for the module.

Examples

iex> MissionControl.help()
%{
  methods: [
    mission: "MissionControl.calculate(integer, list(tuple))",
    mission_stage: "MissionControl.calculate(integer(), atom(), float())"
  ],
  requirenments: [
    weight: [
      type: "Integer. Example: 28801",
      limitations: "Greather than 0"
    ],
    mission_stages: [
      type: "Keyword List[Atom, Float | Atom], Example: [{:launch, 9.807}, {:land, 1.62}]",
    ],
    planets_available: [
      earth: 9.807,
      moon: 1.62,
      mars: 3.711
    ]
  ],
  examples: [
    mission: "MissionControl.calculate(28801, [{:launch, 9.807}, {:land, 1.62}, {:launch, :moon}, {:land, :earth}])",
    mission_stage: "MissionControl.calculate(28801, :land, 9.807) or MissionControl.calculate(28801, :land, :earth)"
  ]
}